+++ crates/amaru-kernel/src/cardano.rs
@@ -74,6 +74,7 @@ pub mod plutus_data;
 pub mod plutus_script;
 pub mod plutus_version;
 pub mod point;
+pub mod pool_certificates;
 pub mod pool_metadata;
 pub mod pool_params;
 pub mod pool_voting_thresholds;