+++ CHANGELOG.md
@@ -63,6 +63,7 @@ Other guiding principles:
 - **amaru-node**: embed best-effort peer snapshots for known networks (for example mainnet, preprod, preview) at build time from cardano-foundation/cardano-configurations; used by default when `--peer-snapshot` is omitted
 - **amaru**: fix the start/restart of a node ([#1095][], [#1098][])
 - **amaru-node**: fix build script to avoid hitting the github API too frequently ([#1108](https://github.com/pragma-org/amaru/pull/1108))
+- **amaru-ledger**: fix the computation of pool updates ([#1109][])
 
 ## [v10.11.20260723](https://github.com/pragma-org/amaru/releases/tag/v10.11.20260723)
 
@@ -252,3 +253,4 @@ Other guiding principles:
 [#1082]: https://github.com/pragma-org/amaru/pull/1082
 [#1095]: https://github.com/pragma-org/amaru/issues/1095
 [#1098]: https://github.com/pragma-org/amaru/pull/1098
+[#1109]: https://github.com/pragma-org/amaru/pull/1109